+++ src/clj/lambdacd/internal/execution.clj
@@ -59,8 +59,7 @@
          immediate-step-result (execute-or-catch step args ctx-with-result-ch)
          processed-async-result (async/<!! processed-async-result-ch)
          complete-step-result (merge  processed-async-result immediate-step-result)
-         processed-final-result (process-final-result complete-step-result ctx)
-         ]
+         processed-final-result (process-final-result complete-step-result ctx)]
      (log/debug (str "executed step " step-id processed-final-result))
      (step-output step-id processed-final-result))))
 
@@ -123,15 +122,21 @@
           new-result
           (recur (cons step-result result) (rest remaining-steps-with-id) new-args))))))
 
-(defn execute-steps
+; TODO: once execute-steps is removed in next release, rename this to execute-steps
+(defn do-execute-steps [steps args ctx & {:keys [step-result-producer is-killed]
+                                           :or   {step-result-producer serial-step-result-producer
+                                                  is-killed            (atom false)}}]
+  (let [base-ctx-with-kill-switch (assoc ctx :is-killed is-killed)
+        step-results (step-result-producer args (contexts-for-steps steps base-ctx-with-kill-switch))]
+    (reduce merge-step-results step-results)))
+
+(defn execute-steps ; DEPRECATED. use core/execute-steps instead (make sure you also use keyword arguments)
   ([steps args ctx]
-    (execute-steps serial-step-result-producer steps args ctx))
+    (do-execute-steps steps args ctx))
   ([step-result-producer steps args ctx]
-   (execute-steps step-result-producer steps args ctx (atom false)))
+   (do-execute-steps steps args ctx :step-result-producer step-result-producer))
    ([step-result-producer steps args ctx is-killed]
-    (let [base-ctx-with-kill-switch (assoc ctx :is-killed is-killed)
-          step-results (step-result-producer args (contexts-for-steps steps base-ctx-with-kill-switch))]
-      (reduce merge-step-results step-results))))
+    (do-execute-steps steps args ctx :step-result-producer step-result-producer :is-killed is-killed)))
 
 (defn- new-base-id-for [step-id]
   (cons 0 step-id))
@@ -145,7 +150,7 @@
 (defn run [pipeline context]
   (let [build-number (pipeline-state/next-build-number context)]
     (let [runnable-pipeline (map eval pipeline)]
-      (execute-steps runnable-pipeline {} (merge context {:step-id [0]
+      (do-execute-steps runnable-pipeline {} (merge context {:step-id [0]
                                                           :build-number build-number})))))
 
 (defn add-result [new-build-number retriggered-build-number initial-ctx [step-id result]]
@@ -172,7 +177,7 @@
           executable-pipeline (map eval pipeline-fragment-to-run)
           new-build-number (pipeline-state/next-build-number context)]
       (duplicate-step-results-not-running-again new-build-number build-number pipeline-history context step-id-to-run)
-      (execute-steps executable-pipeline {} (assoc context :step-id [(dec root-step-number)]
+      (do-execute-steps executable-pipeline {} (assoc context :step-id [(dec root-step-number)]
                                                            :build-number new-build-number)))))
 
 (defn killed? [ctx]