+++ src/clj/lambdacd/steps/control_flow.clj
@@ -1,13 +1,15 @@
 (ns lambdacd.steps.control-flow
   "control flow elements for a pipeline: steps that control the way their child-steps are being run"
   (:require [lambdacd.internal.execution :as execution]
+            [lambdacd.core :as core]
             [clojure.core.async :as async]))
 
 (defn- parallel-step-result-producer [args steps-and-ids]
   (pmap (partial execution/execute-step args) steps-and-ids))
 
 (defn- execute-steps-in-parallel [steps args step-id]
-  (execution/execute-steps parallel-step-result-producer steps args step-id))
+  (core/execute-steps steps args step-id
+                           :step-result-producer parallel-step-result-producer))
 
 (defn ^{:display-type :parallel} in-parallel [& steps]
   (fn [args ctx]
@@ -30,7 +32,9 @@
 (defn ^{:display-type :parallel} either [& steps]
   (fn [args ctx]
     (let [kill-switch (atom false)
-          execute-output (execution/execute-steps step-producer-returning-with-first-successful steps args (execution/new-base-context-for ctx) kill-switch)]
+          execute-output (core/execute-steps steps args (execution/new-base-context-for ctx)
+                                                     :is-killed kill-switch
+                                                     :step-result-producer step-producer-returning-with-first-successful)]
       (reset! kill-switch true)
       (if (= :success (:status execute-output))
         (first (vals (:outputs execute-output)))
@@ -40,4 +44,4 @@
 
 (defn ^{:display-type :container} in-cwd [cwd & steps]
   (fn [args ctx]
-    (execution/execute-steps steps (assoc args :cwd cwd) (execution/new-base-context-for ctx))))
+    (core/execute-steps steps (assoc args :cwd cwd) (execution/new-base-context-for ctx))))